requirements.txt
Django==4.0.4
django-cors-headers==3.12.0
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.1.1
psycopg2==2.9.3
django-storages[azure]